requirements.txt
colorlog==6.7.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
pyqtgraph==0.13.3
Requests==2.31.0
ruamel.base==1.0.0
